Add Windows profile to POM to unpack the JNI
To make a POM that will work across our target platforms, we need
to have it pull and unpack the appropriate set of native binaries
for the current platform. We can use Maven Build Profiles for this
purpose.

The <activation> clause selects for Windows 64 and sets the local
property "jni.classifier", which is used in the native binary
unpacking section. A TODO marks where similar profiles will be
added for Mac and Linux.