Revert the ConvertToOffice sample to use the 18.20.0 release
A bug was discovered in the 18.21.0 release that prevents the
ConvertToOffice sample from working. As a temporary fix we are
reverting this sample to use 18.20.0, which is known to work.

The bug will be fixed in an upcoming release, at which time this
sample will be updated to use that release.