Fix a semgrep report by fixing the code.
Resolve semgrep report by using a list of strings as the argument
to subprocess.Popen, and set shell to False to prevent string intrepetation.

CWE-78

Description
Found subprocess function Popen with shell=True. This is dangerous because this call will spawn the command using a shell process. Doing so propagates current shell settings and variables, which makes it much easier for a malicious actor to execute commands. Use shell=False instead.